// src/models/event-series-entity-connection-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * proxima-nexus-data-plane-api
 * Proxima Nexus Data Plane API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface EventSeriesEntityConnectionDto {
    /**
     * Date/time the connection was created (ISO string)
     */
    'createdAt'?: string;
    /**
     * Date/time the connection was last updated (ISO string)
     */
    'updatedAt'?: string;
    /**
     * Connection type
     */
    'type': EventSeriesEntityConnectionDtoTypeEnum;
    /**
     * Connection state
     */
    'state'?: EventSeriesEntityConnectionDtoStateEnum;
    /**
     * Event series ID
     */
    'seriesId': string;
}

export enum EventSeriesEntityConnectionDtoTypeEnum {
    attendee = 'attendee',
    admin = 'admin',
    owner = 'owner',
    member = 'member',
    friend = 'friend',
    blocked = 'blocked',
    none = 'none'
}
export enum EventSeriesEntityConnectionDtoStateEnum {
    incoming_request = 'incoming_request',
    outgoing_request = 'outgoing_request',
    active = 'active',
    requested = 'requested',
    invited = 'invited'
}